+++ templates/content-page.php
@@ -1,0 +1,4 @@
+<?php while (have_posts()) : the_post(); ?>
+  <?php the_content(); ?>
+  <?php wp_link_pages(array('before' => '<nav class="pagination">', 'after' => '</nav>')); ?>
+<?php endwhile; ?>